#6 Rename submodule (deprecated → v0)
